Add PlayerClass and it's children Warrior and Wizard + tests for them
Created a class "PlayerClass" which is a composition for the Player which has one class.
Created a class "Warrior" which has a few default values and armor requirements
Created a class "Wizard" which has a few default values and armor requirements
Made a test that validates that the player gets the correct "maxHitPoints" as a Wizard and Warrior.